apps/cms/pyproject.toml
[project]
name = "cms"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.7"
dependencies = [
    "boto3>=1.40.17",
    "django==5.*",
    "django-cors-headers>=4.7.0",
    "django-storages>=1.14.6",
    "djangorestframework>=3.16.1",
    "gunicorn>=23.0.0",
    "pillow>=11.3.0",
    "psycopg[binary]>=3.2.9",
    "python-dotenv>=1.1.1",
    "wagtail==7.*",
    "whitenoise>=6.9.0",
]

[dependency-groups]
dev = [
    "model-bakery>=1.20.5",
    "pytest>=8.4.1",
    "pytest-django>=4.11.1",
    "ruff>=0.12.10",
]
